+++ override/modules/contactform/contactform.php
@@ -1,0 +1,10 @@
+<?php
+
+class ContactFormOverride extends Contactform
+{
+    public function sendMessage()
+    {
+        hook::exec('actionContactFormSubmitBefore');
+        parent::sendMessage();
+    }
+}